+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -509,9 +509,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-08-26T07:29:37.805Z"
+    created: Optional[str] = "2026-08-26T09:46:51.830Z"
 
-    updated: Optional[str] = "2026-08-26T07:29:37.805Z"
+    updated: Optional[str] = "2026-08-26T09:46:51.831Z"
 
     updated_by_id: OptionalNullable[str] = UNSET